+++ Plugin/src/main/java/xyz/kyngs/librelogin/bungeecord/BungeeCordBootstrap.java
@@ -11,6 +11,7 @@ import net.md_5.bungee.api.config.ServerInfo;
 import net.md_5.bungee.api.connection.ProxiedPlayer;
 import net.md_5.bungee.api.plugin.Plugin;
 import xyz.kyngs.librelogin.api.provider.LibreLoginProvider;
+import xyz.kyngs.librelogin.api.util.SemanticVersion;
 
 public class BungeeCordBootstrap extends Plugin implements LibreLoginProvider<ProxiedPlayer, ServerInfo> {
 
@@ -40,6 +41,16 @@ public class BungeeCordBootstrap extends Plugin implements LibreLoginProvider<Pr
     @Override
     public BungeeCordLibreLogin getLibreLogin() {
         return libreLogin;
+    }
+
+    @Override
+    public String getVersion() {
+        return getDescription().getVersion();
+    }
+
+    @Override
+    public SemanticVersion getParsedVersion() {
+        return SemanticVersion.parse(getVersion());
     }
 
 }