+++ Plugin/src/main/java/xyz/kyngs/librelogin/common/listener/AuthenticListeners.java
@@ -24,12 +24,8 @@ import java.net.InetAddress;
 import java.sql.Timestamp;
 import java.time.Duration;
 import java.time.LocalDateTime;
-import java.util.regex.Pattern;
 
 public class AuthenticListeners<Plugin extends AuthenticLibreLogin<P, S>, P, S> {
-
-    @SuppressWarnings("RegExpSimplifiable") //I don't believe you
-    private static final Pattern NAME_PATTERN = Pattern.compile("[a-zA-Z0-9_]*");
 
     protected final Plugin plugin;
     protected final PlatformHandle<P, S> platformHandle;
@@ -72,7 +68,7 @@ public class AuthenticListeners<Plugin extends AuthenticLibreLogin<P, S>, P, S> 
     }
 
     protected PreLoginResult onPreLogin(String username, InetAddress address) {
-        if (username.length() > 16 || !NAME_PATTERN.matcher(username).matches()) {
+        if (!plugin.getConfiguration().get(ConfigurationKeys.ALLOWED_NICKNAME_CHARACTERS).matcher(username).matches()) {
             return new PreLoginResult(PreLoginState.DENIED, plugin.getMessages().getMessage("kick-illegal-username"), null);
         }