+++ CHANGELOG.md
@@ -1,3 +1,4 @@
-0.25.0 -
+0.25.0 - RegEx name validation
 
-- 
+- Add RegEx name validation
+- Allow to retrieve the plugin version from LibreLoginProvider